@@ -1,132 +1,318 @@
-﻿using Ardalis.GuardClauses;
-using SmartEnum.Exceptions;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-
-namespace Ardalis.SmartEnum
+﻿namespace Ardalis.SmartEnum
 {
-    /// <summary>
-    /// A base type to use for creating smart enums in .NET.
-    /// TEnum is the type that is inheriting from this class.
-    /// TValue is the type of the enum value, typically int.
-    /// </summary>
-    /// <remarks></remarks>
-    public abstract class SmartEnum<TEnum, TValue> : IEquatable<SmartEnum<TEnum, TValue>>
-        where TEnum : SmartEnum<TEnum, TValue>
-    {
-        private static readonly Lazy<List<TEnum>> _list = new Lazy<List<TEnum>>(ListAllOptions);
+    using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Reflection;
+    using System.Runtime.CompilerServices;
 
-        private static List<TEnum> ListAllOptions()
+    /// <summary>
+    /// A base type to use for creating smart enums with inner value of type <see cref="System.Int32"/>.
+    /// </summary>
+    /// <typeparam name="TEnum">The type that is inheriting from this class.</typeparam>
+    /// <remarks></remarks>
+    public abstract class SmartEnum<TEnum> : 
+        SmartEnum<TEnum, int>
+        where TEnum : SmartEnum<TEnum, int>
+    {
+        protected SmartEnum(string name, int value) :
+            base(name, value)
         {
-            Type t = typeof(TEnum);
-            return t.GetFields(BindingFlags.Public | BindingFlags.Static)
-            .Where(p => t.IsAssignableFrom(p.FieldType))
-            .Select(pi => (TEnum)pi.GetValue(null))
-            .OrderBy(p => p.Name)
-            .ToList();
+        }
+    }
+
+    /// <summary>
+    /// A base type to use for creating smart enums.
+    /// </summary>
+    /// <typeparam name="TEnum">The type that is inheriting from this class.</typeparam>
+    /// <typeparam name="TValue">The type of the inner value.</typeparam>
+    /// <remarks></remarks>
+    public abstract class SmartEnum<TEnum, TValue> : 
+        IEquatable<SmartEnum<TEnum, TValue>>,
+        IComparable<SmartEnum<TEnum, TValue>>
+        where TEnum : SmartEnum<TEnum, TValue>
+        where TValue : struct, IEquatable<TValue>, IComparable<TValue>
+    {
+        static readonly Lazy<Dictionary<string, TEnum>> _fromName = 
+            new Lazy<Dictionary<string, TEnum>>(() => GetAllOptions().ToDictionary(item => item.Name));
+
+        static readonly Lazy<Dictionary<string, TEnum>> _fromNameIgnoreCase = 
+            new Lazy<Dictionary<string, TEnum>>(() => GetAllOptions().ToDictionary(item => item.Name, StringComparer.OrdinalIgnoreCase));
+
+        static readonly Lazy<Dictionary<TValue, TEnum>> _fromValue = 
+            new Lazy<Dictionary<TValue, TEnum>>(() => {
+                // multiple enums with same value are allowed but store only one per value
+                var dictionary = new Dictionary<TValue, TEnum>();
+                foreach (var item in GetAllOptions())
+                {
+                    if (!dictionary.ContainsKey(item._value))
+                        dictionary.Add(item._value, item);
+                }
+                return dictionary;
+            });
+
+        static IEnumerable<TEnum> GetAllOptions()
+        {
+            Type type = typeof(TEnum);
+            foreach (var fieldInfo in type.GetFields(BindingFlags.Public | BindingFlags.Static | BindingFlags.FlattenHierarchy))
+            {
+                object obj = fieldInfo.GetValue(null);
+                if (obj is TEnum value)
+                {
+                    yield return value;
+                }
+            }
         }
 
-        public static List<TEnum> List => _list.Value;
+        /// <summary>
+        /// Gets a collection containing all the instances of <see cref="SmartEnum{TEnum, TValue}"/>.
+        /// </summary>
+        /// <value>A <see cref="IReadOnlyCollection{TEnum}"/> containing all the instances of <see cref="SmartEnum{TEnum, TValue}"/>.</value>
+        /// <remarks>Retrieves all the instances of <see cref="SmartEnum{TEnum, TValue}"/> referenced by public static read-only fields in the current class or its bases.</remarks>
+        public static IReadOnlyCollection<TEnum> List => 
+            _fromName.Value.Values;
 
-        public string Name { get; }
-        public TValue Value { get; protected set; }
+        private readonly string _name;
+        private readonly TValue _value;
+
+        /// <summary>
+        /// Gets the name.
+        /// </summary>
+        /// <value>A <see cref="String"/> that is the name of the <see cref="SmartEnum{TEnum, TValue}"/>.</value>
+        public string Name =>
+            _name;
+
+        /// <summary>
+        /// Gets the value.
+        /// </summary>
+        /// <value>A <typeparamref name="TValue"/> that is the value of the <see cref="SmartEnum{TEnum, TValue}"/>.</value>
+        public TValue Value =>
+            _value;
 
         protected SmartEnum(string name, TValue value)
         {
-            Name = name;
-            Value = value;
+            if (String.IsNullOrEmpty(name))
+                throw new ArgumentException("Argument cannot be null or empty.", nameof(name));
+            
+            _name = name;
+            _value = value;
         }
 
-        protected SmartEnum()
-        {
-            // Required for EF
-        }
+        /// <summary>
+        /// Gets the type of the inner value.
+        /// </summary>
+        /// <value>A <see name="System.Type"/> that is the type of the value of the <see cref="SmartEnum{TEnum, TValue}"/>.</value>
+        public Type GetValueType () => 
+            typeof(TValue);
 
-        public static TEnum FromName(string name)
+        /// <summary>
+        /// Gets the item associated with the specified name.
+        /// </summary>
+        /// <param name="name">The name of the item to get.</param>
+        /// <param name="ignoreCase"><c>true</c> to ignore case during the comparison; otherwise, <c>false</c>.</param>
+        /// <returns>
+        /// The item associated with the specified name. 
+        /// If the specified name is not found, throws a <see cref="KeyNotFoundException"/>.
+        /// </returns>
+        /// <exception cref="ArgumentException"><paramref name="name"/> is <c>null</c>.</exception> 
+        /// <exception cref="SmartEnumNotFoundException"><paramref name="name"/> does not exist.</exception> 
+        /// <seealso cref="SmartEnum{TEnum, TValue}.TryFromName(string, out TEnum)"/>
+        /// <seealso cref="SmartEnum{TEnum, TValue}.TryFromName(string, bool, out TEnum)"/>
+        public static TEnum FromName(string name, bool ignoreCase = false)
         {
-            Guard.Against.NullOrEmpty(name, nameof(name));
-            var result = List.FirstOrDefault(item => string.Equals(item.Name, name, StringComparison.OrdinalIgnoreCase));
-            if (result == null)
+            if (String.IsNullOrEmpty(name))
+                throw new ArgumentException("Argument cannot be null or empty.", nameof(name));
+
+            if (ignoreCase)
+                return FromName(_fromNameIgnoreCase.Value);
+            else
+                return FromName(_fromName.Value);
+
+            TEnum FromName(Dictionary<string, TEnum> dictionary)
             {
-                throw new SmartEnumNotFoundException($"No {typeof(TEnum).Name} with Name \"{name}\" found.");
+                if (!dictionary.TryGetValue(name, out var result))
+                {
+                    throw new SmartEnumNotFoundException($"No {typeof(TEnum).Name} with Name \"{name}\" found.");
+                }
+                return result;
             }
-            return result;
         }
 
+        /// <summary>
+        /// Gets the item associated with the specified name.
+        /// </summary>
+        /// <param name="name">The name of the item to get.</param>
+        /// <param name="result">
+        /// When this method returns, contains the item associated with the specified name, if the key is found; 
+        /// otherwise, <c>null</c>. This parameter is passed uninitialized.</param>
+        /// <returns>
+        /// <c>true</c> if the <see cref="SmartEnum{TEnum, TValue}"/> contains an item with the specified name; otherwise, <c>false</c>.
+        /// </returns>
+        /// <exception cref="ArgumentException"><paramref name="name"/> is <c>null</c>.</exception> 
+        /// <seealso cref="SmartEnum{TEnum, TValue}.FromName(string, bool)"/>
+        /// <seealso cref="SmartEnum{TEnum, TValue}.TryFromName(string, bool, out TEnum)"/>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static bool TryFromName(string name, out TEnum result) =>
+            TryFromName(name, false, out result);
+
+        /// <summary>
+        /// Gets the item associated with the specified name.
+        /// </summary>
+        /// <param name="name">The name of the item to get.</param>
+        /// <param name="ignoreCase"><c>true</c> to ignore case during the comparison; otherwise, <c>false</c>.</param>
+        /// <param name="result">
+        /// When this method returns, contains the item associated with the specified name, if the name is found; 
+        /// otherwise, <c>null</c>. This parameter is passed uninitialized.</param>
+        /// <returns>
+        /// <c>true</c> if the <see cref="SmartEnum{TEnum, TValue}"/> contains an item with the specified name; otherwise, <c>false</c>.
+        /// </returns>
+        /// <exception cref="ArgumentException"><paramref name="name"/> is <c>null</c>.</exception> 
+        /// <seealso cref="SmartEnum{TEnum, TValue}.FromName(string, bool)"/>
+        /// <seealso cref="SmartEnum{TEnum, TValue}.TryFromName(string, out TEnum)"/>
+        public static bool TryFromName(string name, bool ignoreCase, out TEnum result)
+        {
+            if (String.IsNullOrEmpty(name))
+                throw new ArgumentException("Argument cannot be null or empty.", nameof(name));
+
+            if (ignoreCase)
+                return _fromNameIgnoreCase.Value.TryGetValue(name, out result);
+            else
+                return _fromName.Value.TryGetValue(name, out result);
+        }
+
+        /// <summary>
+        /// Gets an item associated with the specified value.
+        /// </summary>
+        /// <param name="value">The value of the item to get.</param>
+        /// <returns>
+        /// The first item found that is associated with the specified value.
+        /// If the specified value is not found, throws a <see cref="KeyNotFoundException"/>.
+        /// </returns>
+        /// <exception cref="SmartEnumNotFoundException"><paramref name="value"/> does not exist.</exception> 
+        /// <seealso cref="SmartEnum{TEnum, TValue}.FromValue(TValue, TEnum)"/>
+        /// <seealso cref="SmartEnum{TEnum, TValue}.TryFromValue(TValue, out TEnum)"/>
         public static TEnum FromValue(TValue value)
         {
-            // Can't use == to compare generics unless we constrain TValue to "class",
-            // which we don't want because then we couldn't use int.
-            var result = List.FirstOrDefault(item => EqualityComparer<TValue>.Default.Equals(item.Value, value));
-            if (result == null)
+            if (!_fromValue.Value.TryGetValue(value, out var result))
             {
                 throw new SmartEnumNotFoundException($"No {typeof(TEnum).Name} with Value {value} found.");
             }
             return result;
         }
 
+        /// <summary>
+        /// Gets an item associated with the specified value.
+        /// </summary>
+        /// <param name="value">The value of the item to get.</param>
+        /// <param name="defaultValue">The value to return when item not found.</param>
+        /// <returns>
+        /// The first item found that is associated with the specified value.
+        /// If the specified value is not found, returns <paramref name="defaultValue"/>.
+        /// </returns>
+        /// <seealso cref="SmartEnum{TEnum, TValue}.FromValue(TValue)"/>
+        /// <seealso cref="SmartEnum{TEnum, TValue}.TryFromValue(TValue, out TEnum)"/>
         public static TEnum FromValue(TValue value, TEnum defaultValue)
         {
-            // Can't use == to compare generics unless we constrain TValue to "class",
-            // which we don't want because then we couldn't use int.
-            var result = List.FirstOrDefault(item => EqualityComparer<TValue>.Default.Equals(item.Value, value));
-            if (result == null)
+            if (!_fromValue.Value.TryGetValue(value, out var result))
             {
-                result = defaultValue;
+                return defaultValue;
             }
             return result;
         }
 
-        public override string ToString() => $"{Name} ({Value})";
-        public override int GetHashCode() => new { Name, Value }.GetHashCode();
-        public override bool Equals(object obj) => Equals(obj as SmartEnum<TEnum, TValue>);
+        /// <summary>
+        /// Gets an item associated with the specified value.
+        /// </summary>
+        /// <param name="value">The value of the item to get.</param>
+        /// <param name="result">
+        /// When this method returns, contains the item associated with the specified value, if the value is found; 
+        /// otherwise, <c>null</c>. This parameter is passed uninitialized.</param>
+        /// <returns>
+        /// <c>true</c> if the <see cref="SmartEnum{TEnum, TValue}"/> contains an item with the specified name; otherwise, <c>false</c>.
+        /// </returns>
+        /// <seealso cref="SmartEnum{TEnum, TValue}.FromValue(TValue)"/>
+        /// <seealso cref="SmartEnum{TEnum, TValue}.FromValue(TValue, TEnum)"/>
+        public static bool TryFromValue(TValue value, out TEnum result) =>
+            _fromValue.Value.TryGetValue(value, out result);
 
-        public bool Equals(SmartEnum<TEnum, TValue> other)
+        public override string ToString() => 
+            _name;
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public override int GetHashCode() =>
+            _value.GetHashCode(); 
+
+        public override bool Equals(object obj) => 
+            (obj is SmartEnum<TEnum, TValue> other) && Equals(other);
+
+        /// <summary>
+        /// Returns a value indicating whether this instance is equal to a specified <see cref="SmartEnum{TEnum, TValue}"/> value.
+        /// </summary>
+        /// <param name="other">An <see cref="SmartEnum{TEnum, TValue}"/> value to compare to this instance.</param>
+        /// <returns><c>true</c> if <paramref name="other"/> has the same value as this instance; otherwise, <c>false</c>.</returns>
+        public virtual bool Equals(SmartEnum<TEnum, TValue> other)
         {
-            if (other is null)
-            {
-                return false;
-            }
-
-            // If the objects share the same reference, return true
-            if (ReferenceEquals(this, other))
-            {
+            // check if same instance
+            if (Object.ReferenceEquals(this, other))
                 return true;
-            }
 
-            // If the runtime types are not the same, return false
-            if (GetType() != other.GetType())
-            {
+            // it's not same instance so 
+            // check if it's not null and is same value
+            if (other is null)
                 return false;
-            }
 
-            // Return true if both name and value match
-            return Name == other.Name && EqualityComparer<TValue>.Default.Equals(Value, other.Value);
+            return _value.Equals(other._value);
         }
 
         public static bool operator ==(SmartEnum<TEnum, TValue> left, SmartEnum<TEnum, TValue> right)
         {
             // Handle null on left side
-            if (left is null)
-            {
-                if (right is null)
-                {
-                    // null == null = true
-                    return true;
-                }
-
-                return false;
-            }
+            if (left is null)   
+                return right is null; // null == null = true
 
             // Equals handles null on right side
             return left.Equals(right);
         }
 
-        public static bool operator !=(SmartEnum<TEnum, TValue> left, SmartEnum<TEnum, TValue> right) => !(left == right);
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static bool operator !=(SmartEnum<TEnum, TValue> left, SmartEnum<TEnum, TValue> right) => 
+            !(left == right);
 
-        public static implicit operator TValue(SmartEnum<TEnum, TValue> smartEnum) => smartEnum.Value;
-        public static explicit operator SmartEnum<TEnum, TValue>(TValue value) => FromValue(value);
+        /// <summary>
+        /// Compares this instance to a specified <see cref="SmartEnum{TEnum, TValue}"/> and returns an indication of their relative values.
+        /// </summary>
+        /// <param name="other">An <see cref="SmartEnum{TEnum, TValue}"/> value to compare to this instance.</param>
+        /// <returns>A signed number indicating the relative values of this instance and <paramref name="other"/>.</returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public virtual int CompareTo(SmartEnum<TEnum, TValue> other) =>
+            _value.CompareTo(other._value);
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static bool operator <(SmartEnum<TEnum, TValue> left, SmartEnum<TEnum, TValue> right) => 
+            left.CompareTo(right) < 0;
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static bool operator <=(SmartEnum<TEnum, TValue> left, SmartEnum<TEnum, TValue> right) => 
+            left.CompareTo(right) <= 0;
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static bool operator >(SmartEnum<TEnum, TValue> left, SmartEnum<TEnum, TValue> right) => 
+            left.CompareTo(right) > 0;
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static bool operator >=(SmartEnum<TEnum, TValue> left, SmartEnum<TEnum, TValue> right) => 
+            left.CompareTo(right) >= 0;
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static implicit operator TValue(SmartEnum<TEnum, TValue> smartEnum) => 
+            smartEnum._value;
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static explicit operator SmartEnum<TEnum, TValue>(TValue value) => 
+            FromValue(value);
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static explicit operator SmartEnum<TEnum, TValue>(TValue? value) => 
+            value.HasValue ? FromValue(value.Value) : null;    
     }
 }
